Fix disastrous merge issue, const issue
So I came across an issue of locality or whatever that's called.
Scope I believe. When I defined it globally by putting it outside
the main function, it wouldn't build. It would return a blank page.
However when I put the categoriesurl in the function itself,
suddenyl everythign was just peachy. Lessons be learned.
Put your variables in a safe place.